Makes CpuStates architecture-specific
