+++ gui/src/vmm/hv/mod.rs
@@ -41,7 +41,7 @@ pub trait Hypervisor: Send + Sync + 'static {
 ///
 /// On AArch64 this represent one Processing Element (PE).
 pub trait Cpu {
-    type States<'a>: CpuStates + 'a
+    type States<'a>: CpuCommit
     where
         Self: 'a;
     type GetStatesErr: Error + Send + 'static;
@@ -59,134 +59,8 @@ pub trait CpuRun: Cpu {
     fn run(&mut self) -> Result<Self::Exit<'_>, Self::RunErr>;
 }
 
-/// States of [`Cpu`].
-pub trait CpuStates {
-    type Err: Error + Send + 'static;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_rax(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_rbx(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_rcx(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_rdx(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_rbp(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_r8(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_r9(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_r10(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_r11(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_r12(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_r13(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_r14(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_r15(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_rdi(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_rdi(&mut self, v: usize);
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_rsi(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_rsi(&mut self, v: usize);
-
-    #[cfg(target_arch = "x86_64")]
-    fn get_rsp(&mut self) -> Result<usize, Self::Err>;
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_rsp(&mut self, v: usize);
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_rip(&mut self, v: usize);
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_cr0(&mut self, v: usize);
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_cr3(&mut self, v: usize);
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_cr4(&mut self, v: usize);
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_efer(&mut self, v: usize);
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_cs(&mut self, ty: u8, dpl: u8, p: bool, l: bool, d: bool);
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_ds(&mut self, p: bool);
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_es(&mut self, p: bool);
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_fs(&mut self, p: bool);
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_gs(&mut self, p: bool);
-
-    #[cfg(target_arch = "x86_64")]
-    fn set_ss(&mut self, p: bool);
-
-    #[cfg(target_arch = "aarch64")]
-    fn set_pstate(&mut self, v: Pstate);
-
-    #[cfg(target_arch = "aarch64")]
-    fn set_sctlr(&mut self, v: Sctlr);
-
-    #[cfg(target_arch = "aarch64")]
-    fn set_mair_el1(&mut self, attrs: u64);
-
-    #[cfg(target_arch = "aarch64")]
-    fn set_tcr(&mut self, v: Tcr);
-
-    /// # Panics
-    /// If `baddr` has non-zero on bit 0 or 48:64.
-    #[cfg(target_arch = "aarch64")]
-    fn set_ttbr0_el1(&mut self, baddr: usize);
-
-    /// # Panics
-    /// If `baddr` has non-zero on bit 0 or 48:64.
-    #[cfg(target_arch = "aarch64")]
-    fn set_ttbr1_el1(&mut self, baddr: usize);
-
-    #[cfg(target_arch = "aarch64")]
-    fn set_sp_el1(&mut self, v: usize);
-
-    #[cfg(target_arch = "aarch64")]
-    fn set_pc(&mut self, v: usize);
-
-    #[cfg(target_arch = "aarch64")]
-    fn set_x0(&mut self, v: usize);
-
-    #[cfg(target_arch = "aarch64")]
-    fn set_x1(&mut self, v: usize);
-
+/// Provides a method to commit [`CpuStates`].
+pub trait CpuCommit: CpuStates {
     fn commit(self) -> Result<(), Self::Err>;
 }
