+++ gui/src/vmm/hv/windows/cpu.rs
@@ -1,4 +1,5 @@
-use crate::vmm::hv::{Cpu, CpuDebug, CpuExit, CpuIo, CpuRun, CpuStates, IoBuf};
+// SPDX-License-Identifier: MIT OR Apache-2.0
+use crate::vmm::hv::{Cpu, CpuCommit, CpuDebug, CpuExit, CpuIo, CpuRun, CpuStates, IoBuf};
 use std::error::Error;
 use std::marker::PhantomData;
 use std::mem::{size_of, zeroed, MaybeUninit};
@@ -291,7 +292,9 @@ impl<'a, 'b> CpuStates for WhpStates<'a, 'b> {
 
         self.dirty = true;
     }
+}
 
+impl<'a, 'b> CpuCommit for WhpStates<'a, 'b> {
     fn commit(self) -> Result<(), Self::Err> {
         if !self.dirty {
             return Ok(());
