+++ gui/src/vmm/x86_64.rs
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
-use super::hv::{Cpu, CpuFeats, CpuStates};
+use super::hv::{Cpu, CpuCommit, CpuFeats, CpuStates};
 use super::ram::RamMap;
 use super::MainCpuError;
